event_server: cyclic resync of nodes
In case of message loss the nodes don't get the sync message,
therefore a cyclic process is added to resend the last tic.

In order to avoid circular updates the sync has to be idenpotent.
Therefore the sync is only takes any action if the new tic is larger
the the last tic known by the node.
The events:merge had to be modified to skip the itc:join action
if the unseen events only contains one merge action. This way
the sync wont generate merge after merge.